requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
configobj==5.0.8
configparser==6.0.1
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
dj-database-url==2.2.0
Django==5.0.4
django-background-tasks==1.2.5
django-compat==1.0.15
django-heroku==0.3.1
django-livereload-server==0.5.1
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.0
filelock==3.13.1
fitz==0.0.1.dev2
google-api-core==2.17.1
google-api-python-client==2.121.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
graphviz==0.20.1
gspread==6.0.2
gunicorn==22.0.0
h11==0.14.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
isodate==0.6.1
kombu==5.3.7
livereload==2.6.3
looseversion==1.3.0
lxml==5.1.0
msgpack==1.0.8
neo4j==5.18.0
networkx==3.2.1
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
pathlib==1.0.1
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
prov==2.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydot==2.0.0
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
pyxnat==1.6.2
rdflib==7.0.0
redis==5.0.3
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
scipy==1.12.0
selenium==4.19.0
service-identity==24.1.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
StrEnum==0.4.15
svgwrite==1.4.3
tornado==6.4
tqdm==4.66.2
traits==6.3.2
trio==0.25.0
trio-websocket==0.11.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
wsproto==1.2.0
zope.interface==6.3
